refactor(Login): Remove unused methods
