+++ lib/common/components/Login.js
@@ -5,9 +5,13 @@ import { useAuth0 } from '@auth0/auth0-react'
 import { useEffect } from 'react'
 import { browserHistory } from 'react-router'
 
-import type { Props } from '../containers/Login'
+type Props = {
+  redirectOnSuccess?: string
+}
 
-export default function Login ({ redirectOnSuccess }: Props) {
+export default function Login ({
+  redirectOnSuccess = window.location.path || '/home'
+}: Props) {
   const auth0 = useAuth0()
   const { isAuthenticated, isLoading } = auth0
   useEffect(() => {
