+++ lib/common/components/StatusModal.js
@@ -3,12 +3,12 @@
 import React, {Component} from 'react'
 import {Modal, Button} from 'react-bootstrap'
 
-import Login from '../containers/Login'
 import * as statusActions from '../../manager/actions/status'
 import * as editorActions from '../../editor/actions/editor'
-
 import type {Props as ContainerProps} from '../containers/CurrentStatusModal'
 import type {ModalStatus} from '../../types/reducers'
+
+import Login from './Login'
 
 type Props = ContainerProps & $Shape<ModalStatus> & {
   clearStatusModal: typeof statusActions.clearStatusModal,
@@ -97,10 +97,6 @@ export default class StatusModal extends Component<Props, State> {
     })
   }
 
-  _onLoginHide = () => {
-    this.setState({ showLogin: false })
-  }
-
   open (props: Props) {
     this.setState({
       showModal: true,
@@ -168,10 +164,7 @@ export default class StatusModal extends Component<Props, State> {
           '/login' on log in button click above.
         */}
         {this.state.showLogin &&
-          <Login
-            onHide={this._onLoginHide}
-            // Ensure redirect occurs after login.
-            redirectOnSuccess={window.location.path} />
+          <Login redirectOnSuccess={window.location.path} />
         }
       </Modal>
     )
