+++ lib/common/containers/App.js
@@ -11,6 +11,7 @@ import { toast } from 'react-toastify'
 import AdminPage from '../../admin/components/AdminPage'
 import MainAlertsViewer from '../../alerts/containers/MainAlertsViewer'
 import ActiveAlertEditor from '../../alerts/containers/ActiveAlertEditor'
+import Login from '../components/Login'
 import PageNotFound from '../components/PageNotFound'
 import { AUTH0_CLIENT_ID, AUTH0_DEFAULT_SCOPE, AUTH0_DOMAIN } from '../constants'
 import ActiveGtfsEditor from '../../editor/containers/ActiveGtfsEditor'
@@ -27,7 +28,6 @@ import type { dispatchFn, getStateFn } from '../../types/reducers'
 
 import ActiveUserRetriever from './ActiveUserRetriever'
 import AppInfoRetriever from './AppInfoRetriever'
-import Login from './Login'
 import wrapComponentInAuthStrategy from './wrapComponentInAuthStrategy'
 
 function loginOptional (ComponentToWrap) {
@@ -90,7 +90,10 @@ export default class App extends React.Component<AppProps> {
         path: '/home(/:projectId)',
         component: authRequired(ActiveUserHomePage)
       },
-      {path: '/login', component: Login},
+      {
+        path: '/login',
+        component: Login
+      },
       {path: '/', component: loginOptional(ActivePublicLandingPage)},
       {
         path: '/project',
