+++ lib/manager/actions/user.js
@@ -22,10 +22,6 @@ const receiveRecentActivity = createAction(
   'RECEIVE_USER_RECENT_ACTIVITY',
   (payload: Array<RecentActivity>) => payload
 )
-const setRedirectOnLogin = createAction(
-  'SET_REDIRECT_ON_LOGIN',
-  (payload: ?string) => payload
-)
 const updatingUserData = createVoidPayloadAction('UPDATING_USER_DATA')
 const userLoggedIn = createAction(
   'USER_LOGGED_IN',
@@ -44,7 +40,6 @@ const voidAction = createVoidPayloadAction('VOID_ACTION')
 
 export type UserActions = ActionType<typeof checkingExistingLogin> |
   ActionType<typeof receiveRecentActivity> |
-  ActionType<typeof setRedirectOnLogin> |
   ActionType<typeof updatingUserData> |
   ActionType<typeof userLoggedIn> |
   ActionType<typeof userLogout> |
@@ -88,18 +83,6 @@ export function getRecentActivity (user: ManagerUserState) {
     return dispatch(secureFetch(url))
       .then(response => response.json())
       .then(activity => dispatch(receiveRecentActivity(activity)))
-  }
-}
-
-/**
- * Login to the application with an optional URL to redirect to on a successful
- * login.
- */
-export function login0 (redirectOnSuccess: ?string) {
-  return function (dispatch: dispatchFn, getState: getStateFn) {
-    dispatch(setRedirectOnLogin(redirectOnSuccess))
-    // Login component exists at login route.
-    browserHistory.push('/login')
   }
 }
