+++ lib/manager/reducers/user.js
@@ -2,7 +2,6 @@
 
 import UserPermissions from '../../common/user/UserPermissions'
 import UserSubscriptions from '../../common/user/UserSubscriptions'
-
 import type {Action} from '../../types/actions'
 import type {ManagerUserState} from '../../types/reducers'
 
@@ -12,14 +11,11 @@ export const defaultState = {
   profile: null,
   permissions: null,
   recentActivity: null,
-  redirectOnSuccess: null,
   subscriptions: null
 }
 
 const user = (state: ManagerUserState = defaultState, action: Action): ManagerUserState => {
   switch (action.type) {
-    case 'SET_REDIRECT_ON_LOGIN':
-      return {...state, redirectOnSuccess: action.payload}
     case 'CHECKING_EXISTING_LOGIN':
       return {...state, isCheckingLogin: true}
     case 'USER_LOGGED_IN':
