+++ lib/types/reducers.js
@@ -586,7 +586,6 @@ export type ManagerUserState = {
   +permissions: ?UserPermissions,
   +profile: ?UserProfile,
   +recentActivity: ?Array<RecentActivity>,
-  +redirectOnSuccess: ?string,
   +subscriptions: ?UserSubscriptions,
   +token: ?string
 }
